requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
gunicorn == 20.1.0
Pillow == 9.5.0
python-dotenv == 1.0.0
mysqlclient == 2.1.1
pip~=23.0.1
pytz~=2023.3
setuptools~=67.8.0
asgiref~=3.7.2
sqlparse~=0.4.4
platformdirs~=3.5.1
PyJWT~=2.7.0
django-ckeditor~=6.5.1